requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
mysql-connector-python==8.0.30
langgraph==0.2.66
pydantic==1.10.2  # or a newer compatible version
